services/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clipador-backend"
version = "0.1.0"
description = "Backend FastAPI + workers para o Clipador"
authors = [{name = "Clipador Team"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "celery[redis]>=5.4.0",
  "asyncpg>=0.29.0",
  "PyJWT>=2.9.0",
  "passlib>=1.7.4",
  "greenlet>=3.0.0",
  "clipador-core",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "mypy",
  "aiosqlite"
]

[tool.hatch.build]
packages = ["src/clipador_backend"]

[tool.uv.sources]
clipador-core = { workspace = true, editable = true }
